requirement.txt
streamlit
langchain
langchain-community
langchain-experimental
python-dotenv
pandas
requests
ollama
torch
transformers
